+++ README.md
@@ -1,4 +1,4 @@
-# Avo project template
+# seedu.avo.Avo project template
 
 This is a project template for a greenfield Java project. It's named _Avo_. Given below are instructions on how to use it.
 
@@ -13,7 +13,7 @@ Prerequisites: JDK 17, update Intellij to the most recent version.
    1. If there are any further prompts, accept the defaults.
 1. Configure the project to use **JDK 17** (not other versions) as explained in [here](https://www.jetbrains.com/help/idea/sdk.html#set-up-jdk).<br>
    In the same dialog, set the **Project language level** field to the `SDK default` option.
-3. After that, locate the `src/main/java/Avo.java` file, right-click it, and choose `Run Avo.main()` (if the code editor is showing compile errors, try restarting the IDE). If the setup is correct, you should see something like the below as the output:
+3. After that, locate the `src/main/java/seedu/avo/seedu.avo.Avo.java` file, right-click it, and choose `Run seedu.avo.Avo.main()` (if the code editor is showing compile errors, try restarting the IDE). If the setup is correct, you should see something like the below as the output:
    ```
-   Hello, I am Avo.
+   Hello, I am seedu.avo.Avo.
    ```
