--- src/main/java/Commands/Command.java
@@ -1,8 +1,0 @@
-package Commands;
-
-import Exceptions.AvoException;
-
-public abstract class Command {
-    public boolean isExit() { return false; }
-    public abstract void execute(String userInput) throws AvoException;
-}
