--- src/main/java/Commands/UnknownCommand.java
@@ -1,10 +1,0 @@
-package Commands;
-
-import Exceptions.AvoException;
-
-public class UnknownCommand extends Command {
-    @Override
-    public void execute(String userInput) throws AvoException {
-        throw new AvoException("OOPS!!! I'm sorry, but I don't know what " + userInput + " means :-(");
-    }
-}
