--- src/main/java/Exceptions/AvoException.java
@@ -1,7 +1,0 @@
-package Exceptions;
-
-public class AvoException extends Exception {
-    public AvoException(String message) {
-        super(message);
-    }
-}
