--- src/main/java/Storage/FileParser.java
@@ -1,7 +1,0 @@
-package Storage;
-
-import Exceptions.AvoException;
-
-public abstract class FileParser<T> {
-    abstract public T parse(String input) throws AvoException;
-}
