--- src/main/java/Tasks/ToDo.java
@@ -1,15 +1,0 @@
-package Tasks;
-
-public class ToDo extends Task {
-    public ToDo(String name) {
-        super(name);
-    }
-    @Override
-    public String formatData() {
-        return String.format("T : %s", super.formatData());
-    }
-    @Override
-    public String toString() {
-        return String.format("[T] %s", super.toString());
-    }
-}
