+++ src/main/java/seedu/avo/commands/CommandName.java
@@ -1,0 +1,20 @@
+package seedu.avo.commands;
+
+public enum CommandName {
+    LIST("list"),
+    TODO("todo"),
+    DEADLINE("deadline"),
+    EVENT("event"),
+    MARK("mark"),
+    UNMARK("unmark"),
+    DELETE("delete"),
+    SEARCH_DATE("on"),
+    EXIT("exit");
+    private final String commandName;
+    CommandName(String commandName) {
+        this.commandName = commandName;
+    }
+    public String getName() {
+        return commandName;
+    }
+}
