+++ src/main/java/seedu/avo/exceptions/AvoException.java
@@ -1,0 +1,7 @@
+package seedu.avo.exceptions;
+
+public class AvoException extends Exception {
+    public AvoException(String message) {
+        super(message);
+    }
+}
