+++ src/main/java/seedu/avo/parser/CommandParser.java
@@ -1,0 +1,15 @@
+package seedu.avo.parser;
+
+import seedu.avo.commands.Command;
+import seedu.avo.commands.CommandManager;
+
+public class CommandParser {
+    private final CommandManager manager;
+    public CommandParser(CommandManager manager) {
+        this.manager = manager;
+    }
+    public Command parse(String input) {
+        String commandStr = input.split(" ")[0];
+        return manager.getCommand(commandStr);
+    }
+}
