+++ src/main/java/seedu/avo/storage/FileParser.java
@@ -1,0 +1,7 @@
+package seedu.avo.storage;
+
+import seedu.avo.exceptions.AvoException;
+
+public abstract class FileParser<T> {
+    abstract public T parse(String input) throws AvoException;
+}
