+++ src/main/java/seedu/avo/tasks/ToDo.java
@@ -1,0 +1,15 @@
+package seedu.avo.tasks;
+
+public class ToDo extends Task {
+    public ToDo(String name) {
+        super(name);
+    }
+    @Override
+    public String formatData() {
+        return String.format("T : %s", super.formatData());
+    }
+    @Override
+    public String toString() {
+        return String.format("[T] %s", super.toString());
+    }
+}
