+++ text-ui-test/runtest.sh
@@ -13,7 +13,7 @@ then
 fi
 
 # compile the code into the bin folder, terminates if error occurred
-if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/*.java
+if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/seedu/avo/*.java
 then
     echo "********** BUILD FAILURE **********"
     exit 1
@@ -31,8 +31,10 @@ diff ACTUAL.TXT EXPECTED-UNIX.TXT
 if [ $? -eq 0 ]
 then
     echo "Test result: PASSED"
+    rm -rf data
     exit 0
 else
     echo "Test result: FAILED"
+    rm -rf data
     exit 1
 fi
